feat-alan-redis : 1. add "RedisUtils.java" class to process redis logic. 2. add "bidHistory" field to Game class.
